✨ feat(MainHeader): add breadcrumbLabel prop for custom titles
- introduce breadcrumbLabel prop to MainHeader component
- allow for page-specific breadcrumb titles (e.g., "Dashboard")
- update Breadcrumb logic to prioritize breadcrumbLabel when available and no active item exists
- ensure correct rendering of different breadcrumb cases based on active item and breadcrumbLabel presence
- add DashboardPage and FlowchartPage components as placeholders for future page-specific logic
- update WorkspaceContext and WorkspaceProvider to include breadcrumbLabel state and setter
- integrate breadcrumbLabel with AppLayout to pass the prop to MainHeader
